logging: Write in variable
Various storages are possible to write logs to
and by now we only support a callback.
Perhaps people want to write logs in a variable (e.g. $_SESSION)
which would need a custom callback doing so.
We help with this by accepting a variable (array)
and use this as reference for spilling log messages into it.